requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
click==8.3.2
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.0
feedparser==6.0.12
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.14.0
multidict==6.7.1
numpy==2.4.4
openai==2.32.0
packaging==26.1
pluggy==1.6.0
propcache==0.4.1
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
Pygments==2.20.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
